Optimize BannerViewPager.

+++ bannerview/src/main/java/com/zhpan/bannerview/indicator/IIndicator.java
@@ -20,7 +20,7 @@ public interface IIndicator extends ViewPager.OnPageChangeListener {
 
     void setSlideMode(@AIndicatorSlideMode int slideStyle);
 
-    void setIndicatorGap(int gap);
+    void setIndicatorGap(float gap);
 
     void setIndicatorWidth(int normalIndicatorWidth, int checkedIndicatorWidth);
